requirements.txt
openai
python-telegram-bot
tiktoken
requests
python-dotenv
asyncio
duckduckgo_search
pytest
pytest-cov
coverage